refactor: update ecmwf opendata
